+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log
 Change log for the Basil Game Engine project.
 
+## 2021-05-04
+### Added
+* Pesto, the Basil Game Engine Editor.
+
+### Changed
+* Updated `README.md`.
+
+
 ## 2021-05-03
 ### Added
 * Basic Framebuffer class.